PD-215 Remove replicated tables
